requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-jsonfield==1.4.0
django-multiselectfield==0.1.12
djangorestframework==3.11.1
djcommon==1.1.1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
validate-email==1.3
whitenoise==5.2.0
zope.interface==5.1.0
